% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMR.R
\docType{data}
\name{AMR}
\alias{AMR}
\title{Active Metabolic Rate: Final Data}
\format{A data frame with 12 rows and 15 variables:
\describe{
  \item{Chamber.No}{The number of a chamber}
  \item{Ind}{ID of an animal}
  \item{Weight}{wet weight of an animal (g)}
  \item{Volume}{the volume of a chamber (mL)}
  \item{Date.Time}{Date and Time of a period (mm/dd/yy hh:mm:ss)}
  \item{Phase}{Phase and period of measurements (e.g. M1, F3)}
  \item{Temp}{Average temperature over a period of measurements (C)}
  \item{Slope.with.BR}{Slope of animal oxygen consumption\cr with slope of background respiration (mgO2/L/h)}
  \item{Slope}{Slope of animal oxygen consumption without background respiration (mgO2/L/h)}
  \item{SE}{Standard error of a slope of animal oxygen consumption\cr without background respiration (mgO2/L/h)}
  \item{R2}{r2 of a slope of animal oxygen consumption\cr without background respiration}
  \item{MR.abs.with.BR}{Absolute AMR with background respiration (mgO2/h)}
  \item{BR}{Percentage rate of background respiration}
  \item{MR.abs}{Absolute AMR (mgO2/h)}
  \item{MR.mass}{Mass-specific AMR (mgO2/kg/h)}
}}
\usage{
AMR
}
\description{
A dataset containing background respiration, absolute and
mass-specific active metabolic rate data obtained by using
the function \code{\link{calculate.MR}}
}
\keyword{datasets}
