% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMR.slope.R
\docType{data}
\name{SMR.slope}
\alias{SMR.slope}
\title{Standard Metabolic Rate: Extracted Slope(s)}
\format{A data frame with 12 rows and 11 variables:
\describe{
  \item{Chamber.No}{The number of a chamber}
  \item{Ind}{ID of an animal}
  \item{Weight}{wet weight of an animal (g)}
  \item{Volume}{the volume of a chamber (mL)}
  \item{Date.Time}{Date and Time of a period (mm/dd/yy hh:mm:ss)}
  \item{Phase}{Phase and period of measurements (e.g. M1, F3)}
  \item{Temp}{Average temperature over a period of measurements (C)}
  \item{Slope.with.BR}{Slope of animal oxygen consumption with slope of\cr background respiration (mgO2/L/h)}
  \item{Slope}{Slope of animal oxygen consumption without background respiration (mgO2/L/h)}
  \item{SE}{Standard error of a slope of animal oxygen consumption without\cr background respiration (mgO2/L/h)}
  \item{R2}{r2 of a slope of animal oxygen consumption without background respiration}
  \item{DO.unit}{the measure unit of DO concentration}
}}
\usage{
SMR.slope
}
\description{
A dataset containing extracted slopes for further SMR calculations and other attributes of standard metabolic rate measurements obtained by using the function \code{\link{extract.slope}}
}
\keyword{datasets}
