% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishResp.R
\docType{package}
\name{FishResp}
\alias{FishResp}
\alias{FishResp-package}
\title{FishResp: An Analytical Tool for Aquatic Respirometry}
\description{
Calculates metabolic rate of fish and other aquatic organisms measured
using an intermittent-flow respirometry approach. The tool is used to
run a set of graphical QC tests of raw respirometry data, correct it for
background respiration and chamber effect, filter and extract target
values of absolute and mass-specific metabolic rate. Experimental design
should include background respiration tests and measuring of one or two
metabolic rate traits. The package allows a user to import raw respirometry
data obtained from 'AutoResp' (LoligoSystems) or other logger software.
}
\author{
Sergey Morozov \email{sergey.morozov@helsinki.fi}
Scott McCairns \email{scott.mccairns@inra.fr}
}
